tg_bot_UI/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.5.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
phonenumbers==8.13.30
python-dotenv==0.21.0
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4